+++ app/apis/member.py
@@ -1,16 +1,16 @@
-# Import all required tools to create rest APIs
 from flask_restx import Namespace, Resource, fields
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from app.apis import MSG
 from app.db import DB
+from app.db.users import UserResource
 from app.db.classes import ClassResource
 from app.db.classes import CLASS_NAME, TRAINER_NAME, CLASS_START_TIME, CLASS_END_TIME, CLASS_DESCRIPTION, CLASS_ROOM_NUMBER, CLASS_CAPACITY, CLASS_USER_IDS
 from datetime import datetime
 from http import HTTPStatus
 from flask import request
 from bson import ObjectId
-
-api = Namespace("classes", description="Endpoint for fitness classlist")
+api = Namespace("member", description="Endpoints for member")
+# VIEW LIST OF ALL CLASSES
 
 _EXAMPLE_CLASS_1 = {
     "Class_name": "Morning_yoga",
@@ -67,6 +67,7 @@ class ClassList(Resource):
     # Possible resposne in Swagger
     @api.response(HTTPStatus.OK, "Success-List of upcoming classes returned")
     def get(self):
+        """Get all classes"""
 
         # Get list of upcoming fitness classes. return all fitness classes regardless of their status
         # Guest/Members can see the complete list of all available options
@@ -113,78 +114,58 @@ class ClassList(Resource):
             })
 
         return result
+# BOOK CLASS FEATURE (FOR MEMBER)
+_UNAUTHORIZED_RESPONSE = api.model(
+    "BookClassUnauthorized", {MSG: fields.String(example="Missing or invalid token")}
+)
 
-    # Get classes the current member is already enrolled in
+@api.route("/<class_id>/book")
+@api.param("class_id", "Class id to book")
+class ClassBooking(Resource):
 
-
-@api.route("/member/enrolled")
-# The resource handler for the class list endpoint. this class handles GET requests for fetching up
-# upcoming classes
-class EnrolledClasses(Resource):
     @jwt_required()
     @api.doc(security="Bearer")
-    @api.marshal_list_with(CLASS_MODEL)
-    # Possible resposne in Swagger
-    @api.response(HTTPStatus.OK, "Success-Your enroled classes returned")
-    @api.response(HTTPStatus.UNAUTHORIZED, "Missing or Invalid JWT token")
-    @api.response(HTTPStatus.BAD_REQUEST, "Invalid user ID format")
-    def get(self):
+    @api.response(
+        HTTPStatus.OK,
+        "Booked",
+        api.model("BookClassOK", {MSG: fields.String(example="Booked successfully")}),
+    )
+    @api.response(
+        HTTPStatus.UNAUTHORIZED,
+        "Unauthorized — valid JWT required",
+        _UNAUTHORIZED_RESPONSE,
+    )
+    @api.response(
+        HTTPStatus.NOT_FOUND,
+        "Not Found",
+        api.model("BookClassNotFound", {MSG: fields.String(example="Class not found")}),
+    )
+    @api.response(
+        HTTPStatus.CONFLICT,
+        "Conflict",
+        api.model("BookClassConflict", {MSG: fields.String(example="Class is full / already booked")}),
+    )
+    def post(self, class_id):
+        """Book a class. Requires a valid member JWT (Bearer token)."""
+        user_id = get_jwt_identity()
 
-        # Get list of all fitness classes the current user is currently enrolled in. return all fitness classes regardless of their status
-        # Only retunrs upcoming classes and helps members see which classes they are already enrolled in
-        # Classes are sorted by their start time
+        user_resource = UserResource()
+        user = user_resource.get_user_by_id(user_id)
+        if user is None:
+            return {MSG: "User not found"}, HTTPStatus.NOT_FOUND
 
-        # Get current user's id form the JWT token
-        current_user_id = get_jwt_identity()
+        class_resource = ClassResource()
+        status = class_resource.add_user_to_class(class_id, user_id)
 
-        if not current_user_id:
-            return {"MSG": "User not found"}, HTTPStatus.UNAUTHORIZED
+        if status == "CLASS_NOT_FOUND":
+            return {MSG: "Class not found"}, HTTPStatus.NOT_FOUND
+        if status == "ALREADY_BOOKED":
+            return {MSG: "User already booked this class"}, HTTPStatus.CONFLICT
+        if status == "CLASS_FULL":
+            return {MSG: "Class is full"}, HTTPStatus.CONFLICT
 
-        try:
-            # convert string id to ObjectID for databse query
+        ok = user_resource.add_class_to_user(user_id, class_id)
+        if not ok:
+            return {MSG: "User not found"}, HTTPStatus.NOT_FOUND
 
-            member_oid = ObjectId(current_user_id)
-
-        except Exception:
-            return {"MSG": "Invalid user ID format"}, HTTPStatus.BAD_REQUEST
-
-        # Find all classes this member is among the users and the class is among upcoming classes
-
-        classes = DB.get_collection("classes")
-        # if collection doesn't exist, return empty collection
-        if classes is None:
-            return []
-
-        now = datetime.now()
-        enrolled = classes.find({
-            CLASS_USER_IDS: member_oid,
-            CLASS_START_TIME: {"$gte": now},
-        }).sort(CLASS_START_TIME, 1)
-
-        # transform each class to match class model
-
-        result = []
-        for c in enrolled:
-            # Calculate class status based on bookings and capacity
-
-            booked = len(c.get(CLASS_USER_IDS, []))
-            capacity = c.get(CLASS_CAPACITY, 0)
-
-            # response for class list
-
-            start_dt = c.get(CLASS_START_TIME)
-            class_data = {
-                "Class_name": c.get(CLASS_NAME, ""),
-                "Trainer_name": c.get(TRAINER_NAME, ""),
-                "Class_date": start_dt.strftime("%Y-%m-%d") if start_dt else "",
-                "Class_start_time": start_dt.strftime("%H:%M") if start_dt else "",
-                "Class_end_time": c.get(CLASS_END_TIME).strftime("%H:%M") if c.get(CLASS_END_TIME) else "",
-                "Class_description": c.get(CLASS_DESCRIPTION, ""),
-                "Class_room_number": c.get(CLASS_ROOM_NUMBER, ""),
-                "Class_capacity": capacity,
-                "Class_status": "Closed" if booked >= capacity else "Open"
-            }
-
-            result.append(class_data)
-
-        return result
+        return {MSG: "Booked successfully"}, HTTPStatus.OK
